+++ CHANGELOG.md
@@ -15,6 +15,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [PR-24]
+
+### Changed
+
+- Each sport's Athlete Unknown is its own URL
+- Future refactoring required
+
 ## [PR-22]
 
 ### Changed